Added environment vars and cd command tracking
Environment variables can now be set using the
export command.  Any variable set via the export
command is propagated to the remote gatekeeper.
The variables only "live" as long as the
immediate globus-job-run session, but are set
for every invocation.

The current working directory is now emulated.
The cd command will check the remote gatekeeper
to see if the path exists.  If it does, then
on subsequent commands, "cd last_set_path" is
executed after the environment export, but
before the command itself.